# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "API для работы с Task manager",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/statuses": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Получить список всех статусов",
                "operationId": "c5b6544699445c4374485174d2ddbd38",
                "responses": {
                    "200": {
                        "description": "Успешно получен список статусов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID статуса",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Название статуса",
                                                "type": "string"
                                            },
                                            "description": {
                                                "description": "Описание статуса",
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "description": "Дата создания статуса",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Дата последнего обновления статуса",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "condition": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID состояния",
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "description": "Название состояния",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Создать новый статус",
                "operationId": "9cd08dad76afea383b8ea130793905c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название статуса",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Описание статуса",
                                        "type": "string"
                                    },
                                    "condition_id": {
                                        "description": "ID состояния",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Статус успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об успешном создании статуса",
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID статуса",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Название статуса",
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "description": "Описание статуса",
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "description": "Дата создания статуса",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "description": "Дата последнего обновления статуса",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "condition": {
                                                    "properties": {
                                                        "id": {
                                                            "description": "ID состояния",
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "description": "Название состояния",
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/statuses/{id}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Показать конкретный статус",
                "operationId": "2ef3272db88fb6af3c53de5272fcbb25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно получен статус",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID статуса",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Название статуса",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Описание статуса",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Дата создания статуса",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего обновления статуса",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "condition": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID состояния",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Название состояния",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Статус не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Обновить статус",
                "operationId": "2f1f091f8ed5dd36829b1fe575d5b9b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название статуса",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Описание статуса",
                                        "type": "string"
                                    },
                                    "condition_id": {
                                        "description": "ID состояния",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID статуса",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Название статуса",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Описание статуса",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Дата создания статуса",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего обновления статуса",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "condition": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID состояния",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Название состояния",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Удалить статус",
                "operationId": "ee24e8f060dc206b17c654c17928b790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус успешно удален"
                    }
                }
            }
        },
...
    },
...
}